requirements.txt
aiogram==3.21.0
python-dotenv==0.9.0
asyncio

